+++ watch/tasks/tracking/from_heatmap.py
@@ -279,7 +279,9 @@ def time_aggregated_polys(
         max_area_sqkm=None,
         # max_area_sqkm=2.25,  # ~1.5x upper tail of truth
         max_area_behavior='drop',
-        thresh_hysteresis=None):
+        thresh_hysteresis=None,
+        polygon_simplify_tolerance=None,
+        ):
     '''
     Track function.
 
@@ -415,6 +417,13 @@ def time_aggregated_polys(
         gids, polys = [], []
 
     polys = [p.to_shapely() for p in polys]
+
+    if polygon_simplify_tolerance is not None:
+        polys = [
+            p.simplify(polygon_simplify_tolerance)
+            for p in polys
+        ]
+
     _TRACKS = gpd.GeoDataFrame(dict(gid=gids, poly=polys), geometry='poly')
     # _TRACKS['track_idx'] = range(len(_TRACKS))
     _TRACKS = _TRACKS.reset_index().rename(columns={'index': 'track_idx'})
@@ -705,6 +714,7 @@ class TimeAggregatedBAS(NewTrackFunction):
     min_area_sqkm: Optional[float] = 0.072
     max_area_sqkm: Optional[float] = 2.25
     max_area_behavior: str = 'drop'
+    polygon_simplify_tolerance: Union[None, float] = None
 
     def create_tracks(self, sub_dset):
         tracks = time_aggregated_polys(
@@ -721,6 +731,7 @@ class TimeAggregatedBAS(NewTrackFunction):
             min_area_sqkm=self.min_area_sqkm,
             max_area_sqkm=self.max_area_sqkm,
             max_area_behavior=self.max_area_behavior,
+            polygon_simplify_tolerance=self.polygon_simplify_tolerance,
         )
         return tracks
 
@@ -750,6 +761,7 @@ class TimeAggregatedSC(NewTrackFunction):
     min_area_sqkm: Optional[float] = None
     max_area_sqkm: Optional[float] = None
     max_area_behavior: str = 'drop'
+    polygon_simplify_tolerance: Union[None, float] = None
 
     def create_tracks(self, sub_dset):
         '''
@@ -790,6 +802,7 @@ class TimeAggregatedSC(NewTrackFunction):
                 min_area_sqkm=self.min_area_sqkm,
                 max_area_sqkm=self.max_area_sqkm,
                 max_area_behavior=self.max_area_behavior,
+                polygon_simplify_tolerance=self.polygon_simplify_tolerance,
             )
 
         return tracks